+++ RELEASE_NOTES.md
@@ -1,3 +1,8 @@
+# v1.1.0
+
+- Added profiling and better error handling (@Cyberboss)
+- Upped default shutdown timeout to 1 hour (@Cyberboss)
+
 # v1.0.0
 
 - Initial release (@Cyberboss)